added basic angular routing and resgister functions

+++ server/src/controllers/charactercontroller.js
@@ -23,38 +23,64 @@ module.exports = {
         }).catch(next);
     },
 
+    getAllCharactersOfUSer(req,res,next){
+        const username = req.user.username;
+
+        UserHelper.getUser(username)
+        .then((userResult) =>{
+            if(userResult === null){
+                res.status(400);
+                res.send({Message: 'User not found'});
+            }else{
+                res.status(200);
+                res.send(userResult.characters);
+            }
+        }).catch(next);
+    },
+
+    getCharacter(req,res,next){
+        CharacterModel.findById({_id: req.params.id})
+        .then((result) =>{
+            if(result === null){
+                res.status(400);
+                res.send({Message: 'Character not found'});
+            }else{
+                res.status(200);
+                res.send(result);
+            }
+        }).catch(next);
+    },
+
     updateCharacter(req,res,next){
         const username = req.user.username;
         const charId = req.params.id;
 
         UserHelper.getUser(username)
         .then((userResult) =>{
-            console.log(userResult.characters + 'From userresult');
-
-            userResult.characters.forEach(element => {
-                if(element._id === charId){
-                    console.log('dit werkt');
-                }
-            });
-
-            //TODO check if charId sits in the list of this user else this user is updateing an char that is not his
-
-
             if(userResult === null){
                 res.status(400);
                 res.send({Message: 'username not found'});
-            }else{
-                CharacterModel.findByIdAndUpdate(charId, {name: req.body.newName, race: req.body.newRace, class:req.body.newClass})
-                .then((result) =>{
-                    if(result === null){
-                        res.status(400);
-                        res.send({Message: 'Character not found'});
-                    }else{
-                        res.status(200);
-                        res.send({Message: 'Character updated'});
-                    }
-                }).catch(next);
-            }
+             }
+
+            userResult.characters.forEach(element => {
+             if(element._id.toString() === charId.toString()){
+                    //character belongs to current user
+                    CharacterModel.findByIdAndUpdate(charId, {name: req.body.newName, race: req.body.newRace, class:req.body.newClass})
+                    .then((result) =>{
+                        if(result === null){
+                            res.status(400);
+                            res.send({Message: 'Character not found'});
+                        }else{
+                            res.status(200);
+                            res.send({Message: 'Character updated'});
+                        }
+                    }).catch(next);
+                } else{
+                    //character doesnt belong to this user
+                    res.status(401);
+                    res.send({Message: 'This resource does not belong to user.'});
+                }
+            });
         })
     },
 
@@ -68,18 +94,23 @@ module.exports = {
                 res.status(400);
                 res.send({Message: 'username not found'});
             }else{
-
-
-                CharacterModel.findByIdAndRemove(charId)
-                .then((result) =>{
-                    if(result === null){
-                        res.status(400);
-                        res.send({Message: 'Character not found'});
+                userResult.characters.forEach(element =>{
+                    if(element._id.toString() === charId.toString()){
+                        CharacterModel.findByIdAndRemove(charId)
+                        .then((result) =>{
+                            if(result === null){
+                                res.status(400);
+                                res.send({Message: 'Character not found'});
+                            }else{
+                                res.status(200);
+                                res.send({Message: 'Character deleted'});
+                            }
+                        }).catch(next);
                     }else{
-                        res.status(200);
-                        res.send({Message: 'Character deleted'});
+                        res.status(401);
+                        res.send({Message: 'This resource does not belong to user'});
                     }
-                }).catch(next);
+                })
             }
         })
     }
